requirements.txt
altair==4.1.0
pandas==1.4.2
numpy==1.20.3
streamlit==1.3.1
pip==21.0.1
gsheetsdb==0.1.13
google-auth==2.3.3